float en num de vol et un commentaire ajouté

+++ src/main/java/com/inti/entity/Voyageur.java
@@ -15,7 +15,10 @@ import javax.persistence.Transient;
 public class Voyageur implements Serializable {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int idVoyageur;
+	private long idVoyageur;
+	
+	//bro what is this
+	
 	private String nomVoyageur;
 	private String prenomVoyageur;
 	private int age;
@@ -23,13 +26,13 @@ public class Voyageur implements Serializable {
 	@Transient
 	private List<Reservation> reservations = new ArrayList<>();
 
-	private int numDeVol;
+	private float numDeVol;
 	
-	public int getNumDeVol() {
+	public float getNumDeVol() {
 		return numDeVol;
 	}
 
-	public void setNumDeVol(int numDeVol) {
+	public void setNumDeVol(float numDeVol) {
 
 		this.numDeVol = numDeVol;
 	}
